+++ packages/soda-tanstack-query/src/utils/createUseQuery.ts
@@ -7,10 +7,7 @@ import {
     UseQueryOptions,
     UseQueryResult,
 } from "@tanstack/react-query"
-import { StrictPartial } from "soda-type"
-
-/** 判断函数的参数是否为必需的 */
-export type IsParamRequired<TFn extends (params: any) => any> = Parameters<TFn> extends [param: any] ? true : false
+import { IsParamRequired, StrictPartial } from "soda-type"
 
 // 要在第一个加一个重载，因为 options 是可选的，如果不加这个重载，options 类型会匹配到 DefinedInitialDataOptions，会误认为 initialData 被传递了，造成返回类型必定不为 undefined