+++ packages/deepsea-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # deepsea-components
 
+## 5.17.21
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.46.1
+    - soda-hooks@6.16.10
+
 ## 5.17.20
 
 ### Patch Changes