+++ packages/sdnext/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-node
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.46.1
+
 ## 0.0.11
 
 ### Patch Changes