+++ packages/sdrr/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-node
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.46.1
+
 ## 0.1.5
 
 ### Patch Changes